feat: upgrade semantic release & other dependencies to latest
Fixes #211

BREAKING CHANGE: Migrated to semantic-release v24. Any breaking
changes from v23 and v24 apply to this github action version
except for Node version requirements. Because this is a docker-based
github action, the version of node in use is defined inside of the
docker image, not by the consuming runner or your code.

deps: if using release.config.js as the name of your config file, it needs to be moved to a .config/ directory or renamed to .releaserc.js to keep it in the root of the repository. see https://github.com/cosmiconfig/cosmiconfig/releases/tag/v9.0.0 for more detail